Move views to separate apps (GH-3)
